third_party/blink/web_tests/external/wpt/css/css-easing/linear-timing-functions-syntax-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['animation-timing-function'] = "linear(0, 1)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0, 1)" but got "linear(0 0%, 1 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(-10, -5, 0, 5, 10)" should set the property value
  assert_equals: serialization should be canonical expected "linear(-10, -5, 0, 5, 10)" but got "linear(-10 0%, -5 25%, 0 50%, 5 75%, 10 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(-10 -10%, -5 -5%, 0, 5, 10)" should set the property value
  assert_equals: serialization should be canonical expected "linear(-10 -10%, -5 -5%, 0, 5, 10)" but got "linear(-10 -10%, -5 -5%, 0 30%, 5 65%, 10 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0 calc(0%), 0 calc(100%))" should set the property value
  assert_equals: serialization should be canonical expected "linear(0 calc(0%), 0 calc(100%))" but got "linear(0 0%, 0 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0 calc(50% - 50%), 0 calc(50% + 50%))" should set the property value
  assert_equals: serialization should be canonical expected "linear(0 calc(0%), 0 calc(100%))" but got "linear(0 0%, 0 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0 calc(50%), 0 100%)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0 calc(50%), 0 100%)" but got "linear(0 50%, 0 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0 0% 50%, 1 50% 100%)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0 0% 50%, 1 50% 100%)" but got "linear(0 0%, 0 50%, 1 50%, 1 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0, 0.5 25% 75%, 1 100% 100%)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0, 0.5 25% 75%, 1 100% 100%)" but got "linear(0 0%, 0.5 25%, 0.5 75%, 1 100%, 1 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0, 1.3, 1, 0.92, 1, 0.99, 1, 0.998, 1 100% 100%)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0, 1.3, 1, 0.92, 1, 0.99, 1, 0.998, 1 100% 100%)" but got "linear(0 0%, 1.3 12.5%, 1 25%, 0.92 37.5%, 1 50%, 0.99 62.5%, 1 75%, 0.998 87.5%, 1 100%, 1 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0, 0 40%, 1, 0.5, 1)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0, 0 40%, 1, 0.5, 1)" but got "linear(0 0%, 0 40%, 1 60%, 0.5 80%, 1 100%)"
[FAIL] e.style['animation-timing-function'] = "linear(0, 1.3, 1, 0.92, 1, 0.99, 1, 1.004, 0.998, 1 100% 100%)" should set the property value
  assert_equals: serialization should be canonical expected "linear(0, 1.3, 1, 0.92, 1, 0.99, 1, 1.004, 0.998, 1 100% 100%)" but got "linear(0 0%, 1.3 11.1111%, 1 22.2222%, 0.92 33.3333%, 1 44.4444%, 0.99 55.5556%, 1 66.6667%, 1.004 77.7778%, 0.998 88.8889%, 1 100%, 1 100%)"
[FAIL] Property animation-timing-function value 'linear(0, 1.3, 1, 0.92, 1, 0.99, 1, 1.004, 0.998, 1 100% 100%)'
  assert_equals: expected "linear(0 0%, 1.3 11.111111%, 1 22.222222%, 0.92 33.333333%, 1 44.444444%, 0.99 55.555556%, 1 66.666667%, 1.004 77.777778%, 0.998 88.888889%, 1 100%, 1 100%)" but got "linear(0 0%, 1.3 11.1111%, 1 22.2222%, 0.92 33.3333%, 1 44.4444%, 0.99 55.5556%, 1 66.6667%, 1.004 77.7778%, 0.998 88.8889%, 1 100%, 1 100%)"
Harness: the test ran to completion.
